Move interface to own file
